chore: clean up buckets types
