+++ packages/buckets/src/index.ts
@@ -1,23 +1,3 @@
-// Exports all API response types for typescript users
-export {
-  ArchiveResponse,
-  ArchiveWatchResponse,
-  CreateResponse,
-  LinksResponse,
-  ListIpfsPathResponse,
-  ListPathResponse,
-  ListResponse,
-  Metadata,
-  PathItem,
-  PullIpfsPathResponse,
-  PullPathResponse,
-  PushPathResponse,
-  RemovePathResponse,
-  RemoveResponse,
-  Root,
-  RootResponse,
-  SetPathResponse,
-} from '@textile/buckets-grpc/api/bucketsd/pb/bucketsd_pb'
-export * from './api'
 export * from './buckets'
+export * from './types'
 export * from './utils'
